move hvcC box into HEVC specific file
