+++ libheif/hevc.cc
@@ -21,7 +21,295 @@
 #include "hevc.h"
 #include "bitstream.h"
 
+#include <cassert>
 #include <cmath>
+#include <cstring>
+#include <iomanip>
+#include <string>
+#include <utility>
+
+Error Box_hvcC::parse(BitstreamRange& range)
+{
+  //parse_full_box_header(range);
+
+  uint8_t byte;
+
+  auto& c = m_configuration; // abbreviation
+
+  c.configuration_version = range.read8();
+  byte = range.read8();
+  c.general_profile_space = (byte >> 6) & 3;
+  c.general_tier_flag = (byte >> 5) & 1;
+  c.general_profile_idc = (byte & 0x1F);
+
+  c.general_profile_compatibility_flags = range.read32();
+
+  for (int i = 0; i < 6; i++) {
+    byte = range.read8();
+
+    for (int b = 0; b < 8; b++) {
+      c.general_constraint_indicator_flags[i * 8 + b] = (byte >> (7 - b)) & 1;
+    }
+  }
+
+  c.general_level_idc = range.read8();
+  c.min_spatial_segmentation_idc = range.read16() & 0x0FFF;
+  c.parallelism_type = range.read8() & 0x03;
+  c.chroma_format = range.read8() & 0x03;
+  c.bit_depth_luma = static_cast<uint8_t>((range.read8() & 0x07) + 8);
+  c.bit_depth_chroma = static_cast<uint8_t>((range.read8() & 0x07) + 8);
+  c.avg_frame_rate = range.read16();
+
+  byte = range.read8();
+  c.constant_frame_rate = (byte >> 6) & 0x03;
+  c.num_temporal_layers = (byte >> 3) & 0x07;
+  c.temporal_id_nested = (byte >> 2) & 1;
+
+  m_length_size = static_cast<uint8_t>((byte & 0x03) + 1);
+
+  int nArrays = range.read8();
+
+  for (int i = 0; i < nArrays && !range.error(); i++) {
+    byte = range.read8();
+
+    NalArray array;
+
+    array.m_array_completeness = (byte >> 6) & 1;
+    array.m_NAL_unit_type = (byte & 0x3F);
+
+    int nUnits = range.read16();
+    for (int u = 0; u < nUnits && !range.error(); u++) {
+
+      std::vector<uint8_t> nal_unit;
+      int size = range.read16();
+      if (!size) {
+        // Ignore empty NAL units.
+        continue;
+      }
+
+      if (range.prepare_read(size)) {
+        nal_unit.resize(size);
+        bool success = range.get_istream()->read((char*) nal_unit.data(), size);
+        assert(success);
+        (void) success;
+      }
+
+      array.m_nal_units.push_back(std::move(nal_unit));
+    }
+
+    m_nal_array.push_back(std::move(array));
+  }
+
+  range.skip_to_end_of_box();
+
+  return range.get_error();
+}
+
+
+std::string Box_hvcC::dump(Indent& indent) const
+{
+  std::ostringstream sstr;
+  sstr << Box::dump(indent);
+
+  const auto& c = m_configuration; // abbreviation
+
+  sstr << indent << "configuration_version: " << ((int) c.configuration_version) << "\n"
+       << indent << "general_profile_space: " << ((int) c.general_profile_space) << "\n"
+       << indent << "general_tier_flag: " << c.general_tier_flag << "\n"
+       << indent << "general_profile_idc: " << ((int) c.general_profile_idc) << "\n";
+
+  sstr << indent << "general_profile_compatibility_flags: ";
+  for (int i = 0; i < 32; i++) {
+    sstr << ((c.general_profile_compatibility_flags >> (31 - i)) & 1);
+    if ((i % 8) == 7) sstr << ' ';
+    else if ((i % 4) == 3) sstr << '.';
+  }
+  sstr << "\n";
+
+  sstr << indent << "general_constraint_indicator_flags: ";
+  int cnt = 0;
+  for (int i = 0; i < configuration::NUM_CONSTRAINT_INDICATOR_FLAGS; i++) {
+    bool b = c.general_constraint_indicator_flags[i];
+
+    sstr << (b ? 1 : 0);
+    cnt++;
+    if ((cnt % 8) == 0)
+      sstr << ' ';
+  }
+  sstr << "\n";
+
+  sstr << indent << "general_level_idc: " << ((int) c.general_level_idc) << "\n"
+       << indent << "min_spatial_segmentation_idc: " << c.min_spatial_segmentation_idc << "\n"
+       << indent << "parallelism_type: " << ((int) c.parallelism_type) << "\n"
+       << indent << "chroma_format: ";
+
+  switch (c.chroma_format) {
+    case 1:
+      sstr << "4:2:0";
+      break;
+    case 2:
+      sstr << "4:2:2";
+      break;
+    case 3:
+      sstr << "4:4:4";
+      break;
+    default:
+      sstr << ((int) c.chroma_format);
+      break;
+  }
+
+  sstr << "\n"
+       << indent << "bit_depth_luma: " << ((int) c.bit_depth_luma) << "\n"
+       << indent << "bit_depth_chroma: " << ((int) c.bit_depth_chroma) << "\n"
+       << indent << "avg_frame_rate: " << c.avg_frame_rate << "\n"
+       << indent << "constant_frame_rate: " << ((int) c.constant_frame_rate) << "\n"
+       << indent << "num_temporal_layers: " << ((int) c.num_temporal_layers) << "\n"
+       << indent << "temporal_id_nested: " << ((int) c.temporal_id_nested) << "\n"
+       << indent << "length_size: " << ((int) m_length_size) << "\n";
+
+  for (const auto& array : m_nal_array) {
+    sstr << indent << "<array>\n";
+
+    indent++;
+    sstr << indent << "array_completeness: " << ((int) array.m_array_completeness) << "\n"
+         << indent << "NAL_unit_type: " << ((int) array.m_NAL_unit_type) << "\n";
+
+    for (const auto& unit : array.m_nal_units) {
+      //sstr << "  unit with " << unit.size() << " bytes of data\n";
+      sstr << indent;
+      for (uint8_t b : unit) {
+        sstr << std::setfill('0') << std::setw(2) << std::hex << ((int) b) << " ";
+      }
+      sstr << "\n";
+      sstr << std::dec;
+    }
+
+    indent--;
+  }
+
+  return sstr.str();
+}
+
+
+bool Box_hvcC::get_headers(std::vector<uint8_t>* dest) const
+{
+  for (const auto& array : m_nal_array) {
+    for (const auto& unit : array.m_nal_units) {
+
+      dest->push_back((unit.size() >> 24) & 0xFF);
+      dest->push_back((unit.size() >> 16) & 0xFF);
+      dest->push_back((unit.size() >> 8) & 0xFF);
+      dest->push_back((unit.size() >> 0) & 0xFF);
+
+      /*
+      dest->push_back(0);
+      dest->push_back(0);
+      dest->push_back(1);
+      */
+
+      dest->insert(dest->end(), unit.begin(), unit.end());
+    }
+  }
+
+  return true;
+}
+
+
+void Box_hvcC::append_nal_data(const std::vector<uint8_t>& nal)
+{
+  NalArray array;
+  array.m_array_completeness = 0;
+  array.m_NAL_unit_type = uint8_t(nal[0] >> 1);
+  array.m_nal_units.push_back(nal);
+
+  m_nal_array.push_back(array);
+}
+
+void Box_hvcC::append_nal_data(const uint8_t* data, size_t size)
+{
+  std::vector<uint8_t> nal;
+  nal.resize(size);
+  memcpy(nal.data(), data, size);
+
+  NalArray array;
+  array.m_array_completeness = 0;
+  array.m_NAL_unit_type = uint8_t(nal[0] >> 1);
+  array.m_nal_units.push_back(std::move(nal));
+
+  m_nal_array.push_back(array);
+}
+
+
+Error Box_hvcC::write(StreamWriter& writer) const
+{
+  size_t box_start = reserve_box_header_space(writer);
+
+  const auto& c = m_configuration; // abbreviation
+
+  writer.write8(c.configuration_version);
+
+  writer.write8((uint8_t) (((c.general_profile_space & 3) << 6) |
+                           ((c.general_tier_flag & 1) << 5) |
+                           (c.general_profile_idc & 0x1F)));
+
+  writer.write32(c.general_profile_compatibility_flags);
+
+  for (int i = 0; i < 6; i++) {
+    uint8_t byte = 0;
+
+    for (int b = 0; b < 8; b++) {
+      if (c.general_constraint_indicator_flags[i * 8 + b]) {
+        byte |= 1;
+      }
+
+      byte = (uint8_t) (byte << 1);
+    }
+
+    writer.write8(byte);
+  }
+
+  writer.write8(c.general_level_idc);
+  writer.write16((c.min_spatial_segmentation_idc & 0x0FFF) | 0xF000);
+  writer.write8(c.parallelism_type | 0xFC);
+  writer.write8(c.chroma_format | 0xFC);
+  writer.write8((uint8_t) ((c.bit_depth_luma - 8) | 0xF8));
+  writer.write8((uint8_t) ((c.bit_depth_chroma - 8) | 0xF8));
+  writer.write16(c.avg_frame_rate);
+
+  writer.write8((uint8_t) (((c.constant_frame_rate & 0x03) << 6) |
+                           ((c.num_temporal_layers & 0x07) << 3) |
+                           ((c.temporal_id_nested & 1) << 2) |
+                           ((m_length_size - 1) & 0x03)));
+
+  size_t nArrays = m_nal_array.size();
+  if (nArrays > 0xFF) {
+    // TODO: error: too many NAL units
+  }
+
+  writer.write8((uint8_t) nArrays);
+
+  for (const NalArray& array : m_nal_array) {
+
+    writer.write8((uint8_t) (((array.m_array_completeness & 1) << 6) |
+                             (array.m_NAL_unit_type & 0x3F)));
+
+    size_t nUnits = array.m_nal_units.size();
+    if (nUnits > 0xFFFF) {
+      // TODO: error: too many NAL units
+    }
+
+    writer.write16((uint16_t) nUnits);
+
+    for (const std::vector<uint8_t>& nal_unit : array.m_nal_units) {
+      writer.write16((uint16_t) nal_unit.size());
+      writer.write(nal_unit);
+    }
+  }
+
+  prepend_header(writer, box_start);
+
+  return Error::Ok;
+}
 
 
 static double read_depth_rep_info_element(BitReader& reader)
